requirements.txt
asgiref==3.3.1
Django==2.2.7
django-configurations==2.2
django-rest-framework==0.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Faker==6.1.1
jmespath==0.10.0
mysqlclient==2.0.3
PyJWT==1.7.1
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2020.5
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
